attention+cache refactoring -- 2
