+++ sarathi/core/datatypes/block.py
@@ -1,9 +1,12 @@
 """Token blocks."""
 
-from typing import List
+from typing import List, Tuple, Union
+from torch import Tensor
+
 
 _BLANK_TOKEN_ID = -1
 
+KVCache = Union[Tuple[Tensor, Tensor], Tensor]
 
 class LogicalTokenBlock:
     """A block that stores a contiguous chunk of tokens from left to right.
