+++ sarathi/model_executor/models/yi.py
@@ -55,7 +55,7 @@ from sarathi.model_executor.weight_utils import (
     load_tensor_parallel_weights,
 )
 from sarathi.transformers_utils.configs.yi import YiConfig
-from sarathi.worker.cache_engine import KVCache
+from sarathi.core.datatypes.block import KVCache
 
 
 class YiMLP(nn.Module):
